Add commands `/start`, `/help`, and `/suggest`
Both `/start` and `/help` do the same: they print the initial help message
to the user.

The `/suggest` command forwards the user's message to the bot's owner.
Strictly speaking, it doesn't "forward" the message. In fact, only the
content of the message (suggestion itself without command) will be sent
as another message.