Fix decryption issues
1. Decryption of base64 strings was broken due to line breakages
produced by the encryption process. The use of a more modern API solves
the problem.
2. Decryption of binary strings was broken due to a stupid mistake with
char/number comparison which always results in 'False'.
3. Long sequences of characters cannot be processed since Telegram gives
the bot only a short part of such messages. This fact was reflected in
the help message.

Note: the map-lambda constructions was replaced with generator
expressions to improve readability of the code.