Implement an inline button for decryption of messages
Here the word "encryption" means all different kinds of message
transformations used by the bot, such as translation to binary or
hexadecimal string, base64 encoding and others that may be introduced in
the future.

Since the maximum length of inline queries is strongly restricted by
Telegram and most of such messages are pretty long, I decided to use
another approach for decryption -- inline buttons with callbacks.

Note that the length of callback data is restricted as well, so we have
to save original messages into an external database and pass their IDs
as data.