Refactor the query handler matching machinery
Implement one of the SOLID principles: the *open-closed principle* which
makes the system truly extensible. To develop a new feature, you no
longer need to see into the entangled and confusing "conditional hell",
responsible for choosing the handle for the query.

Now you only have to extend the 'txtproc.TextProcessor' class and place
your implementation into the 'strconv' module. That's it! All text
processors are loaded by the system automatically when the application
starts.

Localization is also supported. If your classes are called 'FooEncoder'
and 'FooDecoder', you must add strings 'hint_foo_encoder' and
'hint_foo_decoder' to the 'localization.ini' file. They will be used as
the titles for suggested messages.