Relax the exclusivity of text processors
Replace the 'match_exclusive_processor' method with a new
'match_exclusive_processors' one. The new version returns a list of
exclusive handlers. Just like the 'match_simple_processors' method, but
for exclusive processors.

This allows us to have several decoders with the same (or very similar)
patterns for their 'can_process' methods. Note that if the list of
exclusive processors is not empty, all simple text processors will be
discarded as before.