Fix some issues with ABCs and rearrange the tests
During writing the documentation examples I've realized that there are
encoders not able to process any string. So, the 'Universal' mix-in was
taken out from the 'Encoder' ABC.

A new mix-in class was added as well: HTML.

The 'can_process' method was accidentally defined as static. It must be
a class method, of course, to let the implementors use private methods.

The tests turned out to be broken when running from the terminal with
plain 'pytest' command due to wrong imports. Fixed.

Also, the structure of test files was upgraded to make them more
manageable. Note, however, that now you have to manually set the
PYTHONPATH environment variable to the 'app' directory to let 'pytest'
to work properly.