Fix broken build configuration of the Docker image
To install the 'aiotg' library from a git commit, 'git' itself must be
installed inside a Docker image. Since this is not true for the
'python:3.7-alpine' image, the building process was broken after
merging of be28b76e4a151689fb666c3dba4839af64bf34e5.

I forgot to rebuild the image and doesn't catch the issue in time. So,
I think it's better to force rebuilding every time when starting the
application.

Another protector against such failures in the future is to test the
building process of Docker images on Travis CI.

Also, I fucked up and give a wrong name to the 'start-container.sh'
file. It's called so in the README file, but in fact it was named
'start_container.sh' before. Let's fix it.